Adding code to delete employee from json server and list

+++ src/main/java/EmployeePayRollProject/EmployeePayRollService.java
@@ -132,8 +132,19 @@ public class EmployeePayRollService {
 		}
 	}
 
-	public void deleteEmployee(String name) throws CustomSQLException {
+	public void deleteEmployee(String name,String destination) throws CustomSQLException {
+		if(destination.equals("DB"))
 		empPayRollList = employeePayRollDBService.deleteEmployee(name);
+		else
+		{
+			Iterator itr = empPayRollList.iterator(); 
+	        while (itr.hasNext()) 
+	        { 
+	            EmployeePayRoll e = (EmployeePayRoll) itr.next(); 
+	            if(e.name.equals(name))
+	                itr.remove(); 
+	        }
+		}
 	}
 
 	public boolean checkIFDeletedFromList(String name) {
